backend/pyproject.toml
[project]
name = "prescollab-backend"
version = "0.1.0"
description = "PresCollab API (FastAPI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0,<25",
    "alembic>=1.14.0,<2",
    "argon2-cffi>=23.1.0,<24",
    "asyncpg>=0.30.0,<1",
    "email-validator>=2.2.0,<3",
    "fastapi>=0.115.0,<1",
    "pydantic-settings>=2.6.0,<3",
    "python-multipart>=0.0.9,<1",
    "pyjwt[crypto]>=2.10.0,<3",
    "selectolax>=0.3.21,<1",
    "sqlalchemy[asyncio]>=2.0.36,<3",
    "structlog>=24.4.0,<26",
    "uvicorn[standard]>=0.32.0,<1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0,<1",
    "httpx>=0.27.0,<1",
    "polyfactory>=2.0.0,<3",
    "pyright>=1.1.390,<2",
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.24.0,<1",
    "ruff>=0.8.0,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
